+++ messages/ru-RU/app/modules/guard.php
@@ -4,6 +4,29 @@ return [
 
     'It seems that you have not changed your access password for a long time. We recommend that you, periodically, change the password for access to the administrative interface of the site.' => "Похоже, что Вы давно не меняли пароль доступа. Рекомендуем переодически менять пароль доступа к административному интерфейсу сайта.",
 
+
+    'ID' => "ИД",
+    'Client IP' => "IP клиента",
+    'Client Net' => "Сети клиента",
+    'IP range (start)' => "IP диапазон (начало)",
+    'IP range (end)' => "IP диапазон (конец)",
+    'User Agent' => "Польз. агент",
+    'Status' => "Статус",
+    'Reason' => "Причина",
+    'Created' => "Создано",
+    'Updated' => "Обновлено",
+    'Created by' => "Автор создания",
+    'Updated by' => "Автор изменений",
+    'Release' => "Освобождение",
+
+    'Rate limit exceeded.' => "Превышение лимита запросов.",
+    'Overdrive attack detected.' => "Обнаружена атака овердрайва.",
+    'XSS-attack detected.' => "Обнаружена XSS-атака.",
+    'LFI/RFI/RCE attack detected.' => "Обнаружена LFI/RFI/RCE атака.",
+    'PHP-injection detected.' => "Обнаружена попытка PHP-инъекции.",
+    'SQL-injection detected.' => "Обнаружена попытка SQL-инъекции.",
+    'Access denied from security reason.' => "В доступе отказано по соображениям безопасности.",
+
 ];
 
 ?>